+++ packages/component/src/page/trading/desktop/sections/nav/ticker.tsx
@@ -112,14 +112,14 @@ export const Ticker: FC<Props> = (props) => {
                 <div>24h volume</div>
               </div>
             }
-            value={<Numeral rule="human">{data?.["24h_amount"]}</Numeral>}
+            value={<Numeral rule="human" unit={` ${quote}`}>{data?.["24h_amount"]}</Numeral>}
           />
 
           <MemoizedCompnent symbol={props.symbol} />
           <Statistic
             label={"Open interest"}
             value={
-              <Numeral rule="human" unit={quote}>
+              <Numeral rule="human" unit={` ${quote}`}>
                 {new Decimal(data?.open_interest ?? 0)
                   .mul(data?.index_price ?? 0)
                   .toDecimalPlaces(2)