+++ packages/component/src/select/picker.tsx
@@ -1,4 +1,11 @@
-import { forwardRef, SelectHTMLAttributes, useEffect, useMemo, useRef, useState } from "react";
+import {
+  forwardRef,
+  SelectHTMLAttributes,
+  useEffect,
+  useMemo,
+  useRef,
+  useState,
+} from "react";
 import { cva, VariantProps } from "class-variance-authority";
 
 import { SelectOption } from "@/select/select";
@@ -6,7 +13,12 @@ import { cn } from "@/utils/css";
 import { ActionSheet, ActionSheetItem } from "@/sheet";
 import { ArrowIcon } from "@/icon";
 import { useMediaQuery } from "@orderly.network/hooks";
-import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuTrigger } from "@radix-ui/react-dropdown-menu";
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuTrigger,
+} from "@radix-ui/react-dropdown-menu";
 
 const pickerVariants = cva(
   "orderly-flex orderly-flex-row orderly-items-stretch orderly-rounded focus-within:orderly-outline orderly-outline-primary orderly-bg-base-700 orderly-text-base-contrast/50",
@@ -39,10 +51,10 @@ const pickerVariants = cva(
 
 export interface PickerProps
   extends Omit<
-    SelectHTMLAttributes<HTMLSelectElement>,
-    "disabled" | "size" | "color" | "value"
-  >,
-  VariantProps<typeof pickerVariants> {
+      SelectHTMLAttributes<HTMLSelectElement>,
+      "disabled" | "size" | "color" | "value"
+    >,
+    VariantProps<typeof pickerVariants> {
   loading?: boolean;
   label?: string;
   options: SelectOption[];
@@ -58,6 +70,7 @@ export const Picker = forwardRef<PickerRef, PickerProps>(
       color,
       value,
       label,
+      //@ts-ignore
       placeholder,
       className,
       options,