+++ packages/referral/src/hooks/index.ts
@@ -1,2 +1,1 @@
-export { ReferralProvider } from "./referralContext";
-export {}
+export * from "./referralContext";