+++ packages/component/CHANGELOG.md
@@ -1,5 +1,44 @@
 # @orderly.network/react
 
+## 1.4.3
+
+### Patch Changes
+
+- fix bugs;
+- Updated dependencies
+  - @orderly.network/trading-view@1.4.3
+  - @orderly.network/hooks@1.4.3
+  - @orderly.network/types@1.4.3
+  - @orderly.network/utils@1.4.3
+  - @orderly.network/core@1.4.3
+  - @orderly.network/net@1.4.3
+
+## 1.4.2
+
+### Patch Changes
+
+- fix bugs;
+- Updated dependencies
+  - @orderly.network/trading-view@1.4.2
+  - @orderly.network/hooks@1.4.2
+  - @orderly.network/types@1.4.2
+  - @orderly.network/utils@1.4.2
+  - @orderly.network/core@1.4.2
+  - @orderly.network/net@1.4.2
+
+## 1.4.1
+
+### Patch Changes
+
+- fix bugs;
+- Updated dependencies
+  - @orderly.network/hooks@1.4.1
+  - @orderly.network/core@1.4.1
+  - @orderly.network/net@1.4.1
+  - @orderly.network/trading-view@1.4.1
+  - @orderly.network/types@1.4.1
+  - @orderly.network/utils@1.4.1
+
 ## 1.5.0
 
 ### Minor Changes