+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @orderly.network/utils
 
+## 1.4.3
+
+### Patch Changes
+
+- fix bugs;
+
+## 1.4.2
+
+### Patch Changes
+
+- fix bugs;
+
+## 1.4.1
+
+### Patch Changes
+
+- fix bugs;
+
 ## 1.4.0
 
 ### Minor Changes